generator: refactor
